main_test added
