+++ go.mod
@@ -1,0 +1,14 @@
+module go-testify/go-testify
+
+go 1.23.1
+
+require (
+	github.com/magiconair/properties v1.8.9
+	github.com/stretchr/testify v1.10.0
+)
+
+require (
+	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
+)
