+++ main_test.go
@@ -1,0 +1,92 @@
+package main
+
+import (
+	"net/http"
+	"net/http/httptest"
+	"strconv"
+	"strings"
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
+)
+
+var cafeList = map[string][]string{
+	"moscow": []string{"Мир кофе", "Сладкоежка", "Кофе и завтраки", "Сытый студент"},
+}
+
+func mainHandle(w http.ResponseWriter, req *http.Request) {
+	countStr := req.URL.Query().Get("count")
+	if countStr == "" {
+		w.WriteHeader(http.StatusBadRequest)
+		w.Write([]byte("count missing"))
+		return
+	}
+
+	count, err := strconv.Atoi(countStr)
+	if err != nil {
+		w.WriteHeader(http.StatusBadRequest)
+		w.Write([]byte("wrong count value"))
+		return
+	}
+
+	city := req.URL.Query().Get("city")
+
+	cafe, ok := cafeList[city]
+	if !ok {
+		w.WriteHeader(http.StatusBadRequest)
+		w.Write([]byte("wrong city value"))
+		return
+	}
+
+	if count > len(cafe) {
+		count = len(cafe)
+	}
+
+	answer := strings.Join(cafe[:count], ",")
+
+	w.WriteHeader(http.StatusOK)
+	w.Write([]byte(answer))
+}
+
+func TestMainHandlerWhenRequireIsTrue(t *testing.T) {
+
+	req := httptest.NewRequest("GET", "/cafe?count=3&city=moscow", nil) // здесь нужно создать запрос к сервису
+
+	responseRecorder := httptest.NewRecorder()
+	handler := http.HandlerFunc(mainHandle)
+	handler.ServeHTTP(responseRecorder, req)
+
+	body := responseRecorder.Body.String()
+	// здесь нужно добавить необходимые проверки
+	assert.Greater(t, len(body), 0)
+	require.Equal(t, 200, responseRecorder.Code)
+}
+func TestMainHandlerWhenNoSuchCity(t *testing.T) {
+
+	expected := "wrong city value"
+	req := httptest.NewRequest("GET", "/cafe?count=4&city=taganrog", nil)
+
+	responseRecorder := httptest.NewRecorder()
+	handler := http.HandlerFunc(mainHandle)
+	handler.ServeHTTP(responseRecorder, req)
+
+	body := responseRecorder.Body.String()
+
+	assert.Equal(t, expected, body)
+	require.Equal(t, 400, responseRecorder.Code)
+}
+func TestMainHandlerWhenCountMoreThanTotal(t *testing.T) {
+	expectation := 4
+	req := httptest.NewRequest("GET", "/cafe?count=6&city=moscow", nil) // здесь нужно создать запрос к сервису
+
+	responseRecorder := httptest.NewRecorder()
+	handler := http.HandlerFunc(mainHandle)
+	handler.ServeHTTP(responseRecorder, req)
+
+	body := strings.Split(responseRecorder.Body.String(), ",")
+
+	// здесь нужно добавить необходимые проверки
+
+	require.Len(t, body, expectation)
+}
